MAINT: Make room for unittests
