+++ src/zeit/cms/locking/browser/tests/__init__.py
@@ -1,0 +1,1 @@
+# python package
