+++ src/zeit/cms/locking/browser/tests/test_doctest.py
@@ -1,0 +1,10 @@
+import unittest
+import zeit.cms.testing
+
+
+def test_suite():
+    suite = unittest.TestSuite()
+    suite.addTest(zeit.cms.testing.FunctionalDocFileSuite(
+        'README.txt',
+        package='zeit.cms.locking.browser'))
+    return suite
